requirements.txt
absl-py==0.10.0
asgiref==3.2.10
astunparse==1.6.3
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
Django==3.1.2
django-crispy-forms==1.9.2
editdistance==0.5.3
fabio==0.10.2
gast==0.3.3
google-auth==1.22.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.33.1
h5py==2.10.0
idna==2.10
importlib-metadata==2.0.0
Keras-Preprocessing==1.1.2
Markdown==3.3.2
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.4.0.44
opt-einsum==3.3.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
silx==0.13.2
six==1.15.0
sqlparse==0.4.1
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
urllib3==1.25.11
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.3.2
